requirements.txt
flask
opencv-python
deepface
tensorflow
pillow
numpy